Added CheckUserInput()
Function checks the user input when "Enter" is pressed and if present in the correct word array, can call userInputCorrect(), otherwise call userInputWrong(), also accounts for repeated words so you can't keep typing the same word. This is done by resetting the usedWords array when the stop button is pressed.